pyproject.toml
[tool.poetry]
name = "fastapi-news-app"
version = "0.1.0"
description = "A news app built with fastapi and react"
authors = ["sanggusti <gustiowinata0@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
aiofiles = "^0.7.0"
requests = "^2.26.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
